move exercise-list e toggle-row para exercise page
