+++ src/pages/exercises/components/exercises-list/style.css
